# docs/openapi.json
{
...
  "components": {
    "schemas": {
      "ClientMessage": {
        "description": "Client→server frame.",
        "discriminator": {
          "propertyName": "op"
        },
        "oneOf": [
          {
            "description": "Submit an Answer for an outstanding Question. Mismatched shape yields\n`error.validation` (INV-S4) — NOT a state transition.",
            "properties": {
              "answer": {},
              "op": {
                "enum": [
                  "submit_answer"
                ],
                "type": "string"
              },
              "question_id": {
                "type": "string"
              }
            },
            "required": [
              "question_id",
              "answer",
              "op"
            ],
            "type": "object"
          },
          {
            "description": "Test/dev hook so the harness can drive a question→answer round-trip\nend-to-end without a live April loop. Hidden behind the\n`cfg(any(test, feature = \"test-hooks\"))` switch on the server side;\nthe DTO stays here because the OpenAPI surface is one schema.",
            "properties": {
              "kind": {
                "type": "string"
              },
              "op": {
                "enum": [
                  "issue_question"
                ],
                "type": "string"
              },
              "payload": {
                "nullable": true
              },
              "prompt": {
                "type": "string"
              }
            },
            "required": [
              "prompt",
              "kind",
              "op"
            ],
            "type": "object"
          }
        ]
      },
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable SCREAMING_SNAKE_CASE machine string. Once shipped, never reused.",
            "type": "string"
          },
          "details": {
            "description": "Optional structured detail. Omitted on auth failures per INV-A4.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable, lowercase, no trailing period.",
            "type": "string"
          },
          "request_id": {
            "description": "UUIDv7 request correlator echoed in the `X-Request-Id` header.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
      "QuestionEnvelopeDto": {
        "description": "FR-11 / ISC-009 wire shape. The `schema_version` field is **required** on\nthe OpenAPI side because the domain constructor stamps it unconditionally.",
        "properties": {
          "kind": {
            "description": "Question Kind discriminator: `free_text` | `multi_choice` |\n`multi_select` | `boolean` | `file_upload` (DDD §3.3 Value Objects).",
            "type": "string"
          },
          "payload": {
            "description": "Kind-specific payload (e.g. `options` for `multi_choice`). Absent for\n`free_text` and `boolean`. Schema-less by design — the discriminator\ndrives client rendering, not this field's shape.",
            "nullable": true
          },
          "prompt": {
            "type": "string"
          },
          "question_id": {
            "type": "string"
          },
          "schema_version": {
            "description": "Chat-protocol schema version. Always set by the server; clients pin\nto a known version range (DDD MD-5).",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "question_id",
          "prompt",
          "kind"
        ],
        "type": "object"
      },
...
      "ServerMessage": {
        "description": "Server→client frame. All events carry a snake_case discriminator (DSD-621).",
        "discriminator": {
          "propertyName": "event"
        },
        "oneOf": [
          {
            "description": "April issued a Question Envelope. Echoes the full envelope including\n`schema_version` (ISC-009).",
            "properties": {
              "envelope": {
                "$ref": "#/components/schemas/QuestionEnvelopeDto"
              },
              "event": {
                "enum": [
                  "question_issued"
                ],
                "type": "string"
              },
              "session_id": {
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "envelope",
              "event"
            ],
            "type": "object"
          },
          {
            "description": "Answer accepted; echoed back so the GUI can clear its pending state\n(DDD §3.3 Domain Events `question.answered`).",
            "properties": {
              "answer_kind": {
                "type": "string"
              },
              "event": {
                "enum": [
                  "question_answered"
                ],
                "type": "string"
              },
              "question_id": {
                "type": "string"
              },
              "session_id": {
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "question_id",
              "answer_kind",
              "event"
            ],
            "type": "object"
          },
          {
            "description": "POL-S1: timeout fired before an Answer arrived.",
            "properties": {
              "event": {
                "enum": [
                  "question_timed_out"
                ],
                "type": "string"
              },
              "question_id": {
                "type": "string"
              },
              "session_id": {
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "question_id",
              "event"
            ],
            "type": "object"
          },
          {
            "description": "INV-S4 validation failure — kind mismatch or unknown question_id.",
            "properties": {
              "code": {
                "type": "string"
              },
              "event": {
                "enum": [
                  "error_validation"
                ],
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "question_id": {
                "nullable": true,
                "type": "string"
              },
              "session_id": {
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "code",
              "message",
              "event"
            ],
            "type": "object"
          },
          {
            "description": "L4-Session-003 placeholder — issued when a Workflow Command queues\n(POL-S4 first leg). Wired here so the taxonomy is complete in v1's\nOpenAPI schema and downstream GUI codegen sees it.",
            "properties": {
              "command": {
                "type": "string"
              },
              "event": {
                "enum": [
                  "command_accepted"
                ],
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "session_id": {
                "type": "string"
              }
            },
            "required": [
              "session_id",
              "request_id",
              "command",
              "event"
            ],
            "type": "object"
          }
        ]
      },
...
    }
  },
  "info": {
    "description": "Mutagen workflow service. Spec is generated from handler annotations via utoipa per ADR-0002.",
    "license": {
      "name": ""
    },
    "title": "mutagen-service",
    "version": "0.3.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/projects/{project_id}/session": {
      "get": {
        "operationId": "open_session",
        "parameters": [
          {
            "description": "Project UUIDv7.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "DSD-302: chat-protocol schema version. Omit for current. See harness/service/docs/websocket.md for wire shape.",
            "in": "query",
            "name": "schema_version",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientMessage"
              }
            }
          },
          "description": "Client→server WS text-frame shape (informational; HTTP GET carries no body).",
          "required": true
        },
        "responses": {
          "101": {
            "description": "WebSocket upgrade succeeded. See harness/service/docs/websocket.md for the chat-protocol frame schema (utoipa cannot fully express WS shape inline)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported schema_version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Auth middleware rejected the upgrade handshake before the protocol switch."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Another session is already active on this project (ISC-015). Details carry `active_session_id`."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerMessage"
                }
              }
            },
            "description": "Server→client WS text-frame shape (informational; never emitted as an HTTP response)."
          }
        },
        "tags": [
          "session"
        ]
      }
    },
...
  }
}